Simplify and merge emulatorCollection into child-view-container
Now that there's only the one container it can be simpler and in the file directly
