+++ src/child-view-container.js
@@ -1,5 +1,4 @@
 import _ from 'underscore';
-import emulateCollection from './utils/emulate-collection';
 
 // Provide a container to store, retrieve and
 // shut down child views.
@@ -7,7 +6,20 @@ const Container = function() {
   this._init();
 };
 
-emulateCollection(Container.prototype, '_views');
+// Mix in methods from Underscore, for iteration, and other
+// collection related features.
+// Borrowing this code from Backbone.Collection:
+// https://github.com/jashkenas/backbone/blob/1.1.2/backbone.js#L962
+const methods = ['forEach', 'each', 'map', 'find', 'detect', 'filter',
+                  'select', 'reject', 'every', 'all', 'some', 'any', 'include',
+                  'contains', 'invoke', 'toArray', 'first', 'initial', 'rest',
+                  'last', 'without', 'isEmpty', 'pluck', 'reduce', 'partition'];
+
+_.each(methods, function(method) {
+  Container.prototype[method] = function(...args) {
+    return _[method].apply(_, [this._views].concat(args));
+  };
+});
 
 function stringComparator(comparator, view) {
   return view.model && view.model.get(comparator);
